style: 💄 Update component command options
Update the way option values are passed in the component command. For
example if you wanted or didn't want a corresponding test file, the old
syntax would like this: --withTest or --no-withTest. Now with the new
syntax you just do this --withTest=true or --withTest=false this applies
to all the other component options (withStyle, withStory, withLazy).

BREAKING CHANGE: Update the way option values are passed in the component command. For
example if you wanted or didn't want a corresponding test file, the old
syntax looked like this: --withTest or --no-withTest. Now with the new
syntax you just do this --withTest=true or --withTest=false this applies
to all the other component options (withStyle, withStory, withLazy).